Extract LinkedGroupUtils
