+++ common/test/CMakeLists.txt
@@ -89,6 +89,7 @@ set(COMMON_TEST_SOURCE
         "${COMMON_TEST_SOURCE_DIR}/Model/tst_GroupNode.cpp"
         "${COMMON_TEST_SOURCE_DIR}/Model/tst_Issue.cpp"
         "${COMMON_TEST_SOURCE_DIR}/Model/tst_LayerNode.cpp"
+        "${COMMON_TEST_SOURCE_DIR}/Model/tst_LinkedGroupUtils.cpp"
         "${COMMON_TEST_SOURCE_DIR}/Model/tst_ModelUtils.cpp"
         "${COMMON_TEST_SOURCE_DIR}/Model/tst_Node.cpp"
         "${COMMON_TEST_SOURCE_DIR}/Model/tst_NodeCollection.cpp"
